+++ popup/js/search/taxonomySearch.js
@@ -7,60 +7,97 @@
 /**
  * Find bookmarks with given tag
  * Prefers full tag matches, but will also return "starts with" matches
+ * Supports AND search, if multiple tags are chosen, e.g. "#github #pr"
  */
 export function searchTags(searchTerm) {
   const resultDict = {}
+  const intersectionCalc = []
   const tags = getUniqueTags()
-  for (const tagName in tags) {
-    const tag = tags[tagName]
-    if (tagName.toLowerCase().startsWith(searchTerm)) {
-      for (const elIndex of tag) {
-        const el = ext.model.bookmarks[elIndex]
-        const result = {
-          searchScore: tagName === searchTerm ? 1 : 0.8,
-          ...el,
-        }
-        if (!resultDict[el.index]) {
-          resultDict[el.index] = result
-        }
-        // If we already have a match, the one with the higher score wins
-        if (resultDict[el.index].searchScore < result.searchScore) {
-          resultDict[el.index] = result
+
+  const searchTags = searchTerm.split('#').map((el) => el.trim())
+
+  for (const searchTag of searchTags) {
+    if (!searchTag) {
+      continue
+    }
+    const foundResults = []
+    for (const tagName in tags) {
+      const tag = tags[tagName]
+      if (tagName.toLowerCase().startsWith(searchTag)) {
+        for (const elIndex of tag) {
+          const el = ext.model.bookmarks[elIndex]
+          const result = {
+            searchScore: tagName === searchTag ? 1 : 0.8,
+            ...el,
+          }
+
+          if (!resultDict[el.index]) {
+            resultDict[el.index] = result
+          }
+          // If we already have a match, the one with the higher score wins
+          if (resultDict[el.index].searchScore < result.searchScore) {
+            resultDict[el.index] = result
+          }
+          foundResults.push(el.index)
         }
       }
     }
+    intersectionCalc.push(foundResults)
   }
 
-  return Object.values(resultDict)
+  const resultIntersection = intersectionCalc.reduce((a, b) => a.filter((c) => b.includes(c)))
+
+  return resultIntersection.map((resultIndex) => {
+    return resultDict[resultIndex]
+  })
 }
 
 /**
  * Find bookmarks that are part of given folder name
  * Prefers full folder name matches, but will also return "starts with" matches
+ * Supports AND search, if multiple tags are chosen, e.g. "~Sites ~Blogs"
  */
 export function searchFolders(searchTerm) {
   const resultDict = {}
+  const intersectionCalc = []
   const folders = getUniqueFolders()
-  for (const folderName in folders) {
-    const folder = folders[folderName]
-    if (folderName.toLowerCase().startsWith(searchTerm)) {
-      for (const elIndex of folder) {
-        const el = ext.model.bookmarks[elIndex]
-        const result = {
-          searchScore: folderName === searchTerm ? 1 : 0.8,
-          ...el,
-        }
-        if (!resultDict[el.index]) {
-          resultDict[el.index] = result
-        }
-        if (resultDict[el.index].searchScore < result.searchScore) {
-          resultDict[el.index] = result
+
+  const searchFolders = searchTerm.split('~').map((el) => el.trim())
+
+  for (const searchFolder of searchFolders) {
+    if (!searchFolder) {
+      continue
+    }
+    const foundResults = []
+    for (const folderName in folders) {
+      const folder = folders[folderName]
+      if (folderName.toLowerCase().startsWith(searchFolder)) {
+        for (const elIndex of folder) {
+          const el = ext.model.bookmarks[elIndex]
+          const result = {
+            searchScore: folderName === searchFolder ? 1 : 0.8,
+            ...el,
+          }
+          if (!resultDict[el.index]) {
+            resultDict[el.index] = result
+          }
+          if (resultDict[el.index].searchScore < result.searchScore) {
+            resultDict[el.index] = result
+          }
+          foundResults.push(el.index)
         }
       }
     }
+    intersectionCalc.push(foundResults)
   }
 
-  return Object.values(resultDict)
+  const resultIntersection = intersectionCalc.reduce((a, b) => a.filter((c) => b.includes(c)))
+
+  console.info(intersectionCalc, resultIntersection)
+
+  return resultIntersection.map((resultIndex) => {
+    return resultDict[resultIndex]
+  })
 }
 
 /**